OS X support using MacPort - Work in progress
Still not working properly, work in progress...

Mac OS usually uses either Homebrew or MacPort, but in general it should be the same.
Based on this discussion:
https://github.com/JBakamovic/yavide/issues/20

I though of giving it a try, without making a separate installation bash file solely for mac use.

Using MacVim as the GUI and the dependencies installed just fine. 

However, installing it had a problem, I would need to see what is equivalent to yavaide.desktop (xdg-user-dir DESKTOP) for mac.

Anyway, hope it will help someone as a base first step.